Updated POM to build against Spigot, not Bukkit
Note: To build the actual jar, you'll need to clone this repository in
eclipse, add the project as a general project, convert it to a maven
project, then do "Run as" "Maven Build". The default build profile
should generate the jar file for you.

If you don't want to go through that effort, then here's a compiled
version (it has been shown to work on spigot 1.8.8):

https://mega.nz/#!4wIGEbgS!6a5I95mPqlTe69OZIZj6Mm98QqDsaVDbQc6L8kiZi-4